+++ app/feature/auth/components/EmailLoginForm.tsx
@@ -1,0 +1,199 @@
+import React, { useState } from 'react'
+import {
+  View,
+  Text,
+  TextInput,
+  TouchableOpacity,
+  StyleSheet,
+  Alert,
+  ActivityIndicator,
+} from 'react-native'
+
+import { AuthErrorMessage } from '../constants'
+
+import { useAuth } from './AuthContext'
+
+interface EmailLoginFormProps {
+  onSuccess: () => void
+  onToggleMode: () => void
+}
+
+export function EmailLoginForm({
+  onSuccess,
+  onToggleMode,
+}: EmailLoginFormProps) {
+  const { signInWithEmail } = useAuth()
+  const [email, setEmail] = useState('')
+  const [password, setPassword] = useState('')
+  const [loading, setLoading] = useState(false)
+
+  const validateForm = () => {
+    if (!email.trim()) {
+      Alert.alert('오류', '이메일을 입력해주세요.')
+      return false
+    }
+    if (!password.trim()) {
+      Alert.alert('오류', '비밀번호를 입력해주세요.')
+      return false
+    }
+    if (password.length < 6) {
+      Alert.alert('오류', '비밀번호는 6자 이상이어야 합니다.')
+      return false
+    }
+    return true
+  }
+
+  const handleLogin = async () => {
+    if (!validateForm()) return
+    setLoading(true)
+
+    try {
+      await signInWithEmail(email.trim(), password)
+      onSuccess()
+    } catch (error) {
+      console.log(error)
+      // get error message
+      const errorMessage = (error as Error).message
+      console.log(errorMessage)
+      if (errorMessage === AuthErrorMessage.EMAIL_NOT_CONFIRMED) {
+        Alert.alert('로그인 실패', '이메일 인증을 완료해주세요.')
+      } else {
+        Alert.alert('로그인 실패', '아이디와 비밀번호를 확인해주세요.')
+      }
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>이메일 로그인</Text>
+
+      <View style={styles.form}>
+        <View style={styles.inputGroup}>
+          <Text style={styles.label}>이메일</Text>
+          <TextInput
+            style={styles.input}
+            value={email}
+            onChangeText={setEmail}
+            placeholder="이메일을 입력하세요"
+            keyboardType="email-address"
+            autoCapitalize="none"
+            autoCorrect={false}
+            editable={!loading}
+          />
+        </View>
+
+        <View style={styles.inputGroup}>
+          <Text style={styles.label}>비밀번호</Text>
+          <TextInput
+            style={styles.input}
+            value={password}
+            onChangeText={setPassword}
+            placeholder="비밀번호를 입력하세요 (6자 이상)"
+            secureTextEntry
+            autoCapitalize="none"
+            autoCorrect={false}
+            editable={!loading}
+          />
+        </View>
+
+        <TouchableOpacity
+          style={[styles.submitButton, loading && styles.disabledButton]}
+          onPress={handleLogin}
+          disabled={loading}
+        >
+          {loading ? (
+            <ActivityIndicator size="small" color="#FFF" />
+          ) : (
+            <Text style={styles.submitButtonText}>로그인</Text>
+          )}
+        </TouchableOpacity>
+
+        <TouchableOpacity
+          style={styles.toggleButton}
+          onPress={onToggleMode}
+          disabled={loading}
+        >
+          <Text style={styles.toggleButtonText}>
+            계정이 없으신가요? 회원가입
+          </Text>
+        </TouchableOpacity>
+      </View>
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  title: {
+    fontSize: 28,
+    fontWeight: 'bold',
+    color: '#333',
+    textAlign: 'center',
+    marginBottom: 40,
+  },
+  form: {
+    flex: 1,
+  },
+  inputGroup: {
+    marginBottom: 20,
+  },
+  label: {
+    fontSize: 16,
+    fontWeight: '600',
+    color: '#333',
+    marginBottom: 8,
+  },
+  input: {
+    backgroundColor: '#FFF',
+    borderRadius: 10,
+    paddingHorizontal: 15,
+    paddingVertical: 12,
+    fontSize: 16,
+    borderWidth: 1,
+    borderColor: '#E0E0E0',
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 1,
+    },
+    shadowOpacity: 0.05,
+    shadowRadius: 2,
+    elevation: 2,
+  },
+  submitButton: {
+    backgroundColor: '#007AFF',
+    borderRadius: 10,
+    paddingVertical: 15,
+    alignItems: 'center',
+    marginTop: 20,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  disabledButton: {
+    backgroundColor: '#A0A0A0',
+  },
+  submitButtonText: {
+    color: '#FFF',
+    fontSize: 18,
+    fontWeight: '600',
+  },
+  toggleButton: {
+    marginTop: 20,
+    alignItems: 'center',
+  },
+  toggleButtonText: {
+    fontSize: 14,
+    color: '#007AFF',
+    textDecorationLine: 'underline',
+  },
+})
