+++ app/feature/auth/constants/index.ts
@@ -1,0 +1,4 @@
+export enum AuthErrorMessage {
+  INVALID_CREDENTIALS = 'Invalid credentials',
+  EMAIL_NOT_CONFIRMED = 'Email not confirmed',
+}
