+++ app/feature/auth/index.ts
@@ -1,2 +1,3 @@
 export * from "./components";
 export * from "./lib";
+export * from "./constants";
